+++ src/pptx/oxml/__init__.py
@@ -93,6 +93,7 @@ from pptx.oxml.chart.axis import (  # noqa: E402
     CT_CrossBetween,
     CT_Crosses,
     CT_DateAx,
+    CT_LblAlgn,
     CT_LblOffset,
     CT_Orientation,
     CT_Scaling,
@@ -107,6 +108,7 @@ register_element_cls("c:catAx", CT_CatAx)
 register_element_cls("c:crossBetween", CT_CrossBetween)
 register_element_cls("c:crosses", CT_Crosses)
 register_element_cls("c:dateAx", CT_DateAx)
+register_element_cls("c:lblAlgn", CT_LblAlgn)
 register_element_cls("c:lblOffset", CT_LblOffset)
 register_element_cls("c:majorGridlines", CT_ChartLines)
 register_element_cls("c:majorTickMark", CT_TickMark)